// src/app/shared/types/api-types/bire-configuration-dto.interface.ts
export interface BireConfigurationDTO {
  configurationCode?: string;
  familyCode?: string;
  structureType?: string;
  variantCode?: string;
  chapter?: string;
  sheet?: string;
  marks?: string;
  section?: string;
  subject?: string;
  versionNumber?: string;
  configurationName?: string;
  vintageConfigDate?: Date;
  statusState?: string;
  statusUser?: string;
  statusDate?: Date;
}