// src/app/shared/types/api-types/bire-configuration-pack-dto.interface.ts
export interface BireConfigurationPackDTO {
  configurationCode?: string;
  familyCode?: string;
  structureType?: string;
  variantCode?: string;
  chapter?: string;
  sheet?: string;
  marks?: string;
  section?: string;
  subject?: string;
  packBaseOptionCode?: string;
  statusState?: string;
  statusUser?: string;
  statusDate?: Date;
}