// src/app/shared/types/api-types/bire-item-version-dto.interface.ts
export interface BireItemVersionDTO {
  familyCode?: string;
  structureType?: string;
  variantCode?: string;
  chapter?: string;
  sheet?: string;
  marks?: string;
  section?: string;
  subject?: string;
  versionNumber?: string;
  statusState?: string;
  statusUser?: string;
  statusDate?: Date;
}