// src/app/shared/types/api-types/bire-version-pn-dto.interface.ts
export interface BireVersionPnDTO {
  familyCode?: string;
  structureType?: string;
  variantCode?: string;
  chapter?: string;
  sheet?: string;
  marks?: string;
  section?: string;
  subject?: string;
  versionNumber?: string;
  pnCode?: string;
  isnAta?: string;
  repairableState?: string;
  pnEffectivity?: string;
  ar?: boolean;
  remarks?: string;
  pnQuantityTci?: string;
  isnAecma?: string;
  pnQuantityMpl?: string;
  procurability?: boolean;
  pnAlternativeType?: string;
  rf?: boolean;
  usedWith?: string;
  configurationRule?: string;
  applicability?: string;
  statusState?: string;
  statusUser?: string;
  statusDate?: Date;
}